home page

+++ app/src/main/java/com/opsc/powerpath/Data/Models/WorkoutProgress.kt
@@ -1,5 +1,5 @@
 package com.opsc.powerpath.Data.Models
 
 data class WorkoutProgress(
-    val sets: Int
+    val sets: Float
 )
